Move dataset classes into separate directory and object library
