+++ src/tests/NeXusDataset/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(NeXusDataset_SRC
+        ExtensibleDatasetTests.cpp
+        NeXusDatasetTests.cpp
+        )
+
+add_library(NeXusDatasetTests OBJECT ${NeXusDataset_SRC})
+target_include_directories(NeXusDatasetTests PRIVATE .. ../../)
