+++ src/writer_modules/f142_test/f142_TestWriter.h
@@ -11,7 +11,7 @@
 
 #include "FlatbufferMessage.h"
 #include "WriterModuleBase.h"
-#include <NeXusDataset.h>
+#include <NeXusDataset/NeXusDataset.h>
 
 namespace WriterModule {
 namespace f142 {
